LLM/requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
tf-keras
faiss-cpu
langchain_chroma
beautifulsoup4
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
streamlit